# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Course Enrollment System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  price: number;
  content: string;
  thumbnailUrl?: string;
  createdAt: string;
}

export interface CreateCourseRequest {
  title: string;
  description: string;
  /** @maximum 500 */
  price: number;
  content: string;
  thumbnailUrl?: string;
}

export type EnrollmentPaymentMethod =
  (typeof EnrollmentPaymentMethod)[keyof typeof EnrollmentPaymentMethod];

export const EnrollmentPaymentMethod = {
  upi: "upi",
  credit_card: "credit_card",
  net_banking: "net_banking",
} as const;

export type EnrollmentPaymentStatus =
  (typeof EnrollmentPaymentStatus)[keyof typeof EnrollmentPaymentStatus];

export const EnrollmentPaymentStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Enrollment {
  id: number;
  userId: number;
  courseId: number;
  paymentMethod: EnrollmentPaymentMethod;
  paymentStatus: EnrollmentPaymentStatus;
  amountPaid: number;
  transactionId: string;
  enrolledAt: string;
}

export type EnrollmentWithCoursePaymentMethod =
  (typeof EnrollmentWithCoursePaymentMethod)[keyof typeof EnrollmentWithCoursePaymentMethod];

export const EnrollmentWithCoursePaymentMethod = {
  upi: "upi",
  credit_card: "credit_card",
  net_banking: "net_banking",
} as const;

export type EnrollmentWithCoursePaymentStatus =
  (typeof EnrollmentWithCoursePaymentStatus)[keyof typeof EnrollmentWithCoursePaymentStatus];

export const EnrollmentWithCoursePaymentStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface EnrollmentWithCourse {
  id: number;
  courseId: number;
  paymentMethod: EnrollmentWithCoursePaymentMethod;
  paymentStatus: EnrollmentWithCoursePaymentStatus;
  amountPaid: number;
  transactionId: string;
  enrolledAt: string;
  course: Course;
}

export type PaymentRequestPaymentMethod =
  (typeof PaymentRequestPaymentMethod)[keyof typeof PaymentRequestPaymentMethod];

export const PaymentRequestPaymentMethod = {
  upi: "upi",
  credit_card: "credit_card",
  net_banking: "net_banking",
  qr_code: "qr_code",
} as const;

export type PaymentRequestPaymentDetails = { [key: string]: unknown };

export interface PaymentRequest {
  courseId: number;
  paymentMethod: PaymentRequestPaymentMethod;
  paymentDetails: PaymentRequestPaymentDetails;
}

export interface PaymentResponse {
  success: boolean;
  transactionId: string;
  enrollment?: Enrollment | null;
  message: string;
}

export interface UpdateProfileRequest {
  name: string;
}

export interface ChatbotRequest {
  message: string;
}

export interface ChatbotResponse {
  reply: string;
}