# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Online Course Enrollment System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user (student or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /enrollments:
    get:
      operationId: getMyEnrollments
      tags: [enrollments]
      summary: Get current student's enrollments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of enrollments with course data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrollmentWithCourse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /enrollments/pay:
    post:
      operationId: processPayment
      tags: [enrollments]
      summary: Process payment and enroll in a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "201":
          description: Payment successful and enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          description: Bad request or already enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        content:
          type: string
        thumbnailUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - price
        - content
        - createdAt
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        paymentMethod:
          type: string
          enum: [upi, credit_card, net_banking]
        paymentStatus:
          type: string
          enum: [pending, completed, failed]
        amountPaid:
          type: number
        transactionId:
          type: string
        enrolledAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - courseId
        - paymentMethod
        - paymentStatus
        - amountPaid
        - transactionId
        - enrolledAt
    EnrollmentWithCourse:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        paymentMethod:
          type: string
          enum: [upi, credit_card, net_banking]
        paymentStatus:
          type: string
          enum: [pending, completed, failed]
        amountPaid:
          type: number
        transactionId:
          type: string
        enrolledAt:
          type: string
          format: date-time
        course:
          $ref: "#/components/schemas/Course"
      required:
        - id
        - courseId
        - paymentMethod
        - paymentStatus
        - amountPaid
        - transactionId
        - enrolledAt
        - course
    PaymentRequest:
      type: object
      properties:
        courseId:
          type: integer
        paymentMethod:
          type: string
          enum: [upi, credit_card, net_banking, qr_code]
        paymentDetails:
          type: object
          additionalProperties: true
      required:
        - courseId
        - paymentMethod
        - paymentDetails
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        enrollment:
          $ref: "#/components/schemas/Enrollment"
          nullable: true
        message:
          type: string
      required:
        - success
        - transactionId
        - message
...