# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Online Course Enrollment System API
 * OpenAPI spec version: 0.1.0
 */
import type { Enrollment } from "./enrollment";

export interface PaymentResponse {
  success: boolean;
  transactionId: string;
  enrollment?: Enrollment | null;
  message: string;
}